added basic auth to puts, posts, and deletes.. updates tests accordingly

+++ lrs/tests/ActivityProfileTests.py
@@ -8,10 +8,8 @@ import hashlib
 import urllib
 from os import path
 import sys
-
-_DIR = path.abspath(path.dirname(__file__))
-sys.path.append(path.abspath(path.join(_DIR,"../objects")))
 from lrs.objects import Activity
+import base64
 
 #TODO: delete profiles that are being stored in /var/www/adllrs/media/activity profiles
 class ActivityProfileTests(TestCase):
@@ -26,6 +24,11 @@ class ActivityProfileTests(TestCase):
     otherprofileId1 = "http://profile.test.id/other/1"
 
     def setUp(self):
+        self.username = "tester"
+        self.password = "test"
+        self.auth = "Basic %s" % base64.b64encode("%s:%s" % (self.username, self.password))
+        form = {'username':self.username,'password':self.password,'password2':self.password}
+        response = self.client.post(reverse(views.register),form)
 
         self.act1 = Activity.Activity(json.dumps({'objectType':'Activity', 'id': self.test_activityId1}))
         self.act2 = Activity.Activity(json.dumps({'objectType':'Activity', 'id': self.test_activityId2}))
@@ -35,35 +38,35 @@ class ActivityProfileTests(TestCase):
         self.testparams1 = {"profileId": self.testprofileId1, "activityId": self.test_activityId1}
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams1))
         self.testprofile1 = {"test":"put profile 1","obj":{"activity":"test"}}
-        self.put1 = self.client.put(path, self.testprofile1, content_type=self.content_type)
+        self.put1 = self.client.put(path, self.testprofile1, content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
         
         self.testparams2 = {"profileId": self.testprofileId2, "activityId": self.test_activityId2}
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams2))
         self.testprofile2 = {"test":"put profile 2","obj":{"activity":"test"}}
-        self.put2 = self.client.put(path, self.testprofile2, content_type=self.content_type)
+        self.put2 = self.client.put(path, self.testprofile2, content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
 
         self.testparams3 = {"profileId": self.testprofileId3, "activityId": self.test_activityId3}
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams3))
         self.testprofile3 = {"test":"put profile 3","obj":{"activity":"test"}}
-        self.put3 = self.client.put(path, self.testprofile3, content_type=self.content_type)
+        self.put3 = self.client.put(path, self.testprofile3, content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
 
         self.testparams4 = {"profileId": self.otherprofileId1, "activityId": self.other_activityId}
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams4))
         self.otherprofile1 = {"test":"put profile other","obj":{"activity":"other"}}
-        self.put4 = self.client.put(path, self.otherprofile1, content_type=self.content_type)
+        self.put4 = self.client.put(path, self.otherprofile1, content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
 
         self.testparams5 = {"profileId": self.otherprofileId1, "activityId": self.test_activityId1}
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams5))
         self.anotherprofile1 = {"test":"put another profile 1","obj":{"activity":"other"}}
-        self.put5 = self.client.put(path, self.anotherprofile1, content_type=self.content_type)
+        self.put5 = self.client.put(path, self.anotherprofile1, content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
 
         
     def tearDown(self):
-        self.client.delete(reverse(views.activity_profile), self.testparams1)
-        self.client.delete(reverse(views.activity_profile), self.testparams2)
-        self.client.delete(reverse(views.activity_profile), self.testparams3)
-        self.client.delete(reverse(views.activity_profile), self.testparams4)    
-        self.client.delete(reverse(views.activity_profile), self.testparams5)    
+        self.client.delete(reverse(views.activity_profile), self.testparams1, HTTP_AUTHORIZATION=self.auth)
+        self.client.delete(reverse(views.activity_profile), self.testparams2, HTTP_AUTHORIZATION=self.auth)
+        self.client.delete(reverse(views.activity_profile), self.testparams3, HTTP_AUTHORIZATION=self.auth)
+        self.client.delete(reverse(views.activity_profile), self.testparams4, HTTP_AUTHORIZATION=self.auth)    
+        self.client.delete(reverse(views.activity_profile), self.testparams5, HTTP_AUTHORIZATION=self.auth)    
     
     def test_put(self):
         #Test the puts
@@ -97,23 +100,23 @@ class ActivityProfileTests(TestCase):
 
         
     def test_put_no_params(self):
-        put = self.client.put(reverse(views.activity_profile) ,content_type=self.content_type)
+        put = self.client.put(reverse(views.activity_profile) ,content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
         self.assertEquals(put.content, 'Error -- activity_profile - method = PUT, but activityId parameter missing..')
 
     def test_put_no_activityId(self):
-        put = self.client.put(reverse(views.activity_profile), {'profileId':'10'},content_type=self.content_type)
+        put = self.client.put(reverse(views.activity_profile), {'profileId':'10'},content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
         self.assertEquals(put.content, 'Error -- activity_profile - method = PUT, but activityId parameter missing..')
 
     def test_put_no_profileId(self):
         testparams = {'activityId':'act'}
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(testparams))
-        put = self.client.put(path, content_type=self.content_type)
+        put = self.client.put(path, content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
         self.assertEquals(put.content, 'Error -- activity_profile - method = PUT, but profileId parameter missing..')
 
     def test_put_etag_missing_on_change(self):
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams1))
         profile = {"test":"error - trying to put new profile w/o etag header","obj":{"activity":"test"}}
-        response = self.client.put(path, profile, content_type=self.content_type)
+        response = self.client.put(path, profile, content_type=self.content_type, HTTP_AUTHORIZATION=self.auth)
         self.assertEqual(response.status_code, 409)
         self.assertIn('If-Match and If-None-Match headers were missing', response.content)
         
@@ -125,7 +128,7 @@ class ActivityProfileTests(TestCase):
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams1))
         profile = {"test":"good - trying to put new profile w/ etag header","obj":{"activity":"test"}}
         thehash = '"%s"' % hashlib.sha1('%s' % self.testprofile1).hexdigest()
-        response = self.client.put(path, profile, content_type=self.content_type, if_match=thehash)
+        response = self.client.put(path, profile, content_type=self.content_type, if_match=thehash, HTTP_AUTHORIZATION=self.auth)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.content, 'Success -- activity profile - method = PUT - profileId = %s' % self.testprofileId1)
 
@@ -137,7 +140,7 @@ class ActivityProfileTests(TestCase):
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams1))
         profile = {"test":"error - trying to put new profile w/ wrong etag value","obj":{"activity":"test"}}
         thehash = '"%s"' % hashlib.sha1('%s' % 'wrong hash').hexdigest()
-        response = self.client.put(path, profile, content_type=self.content_type, if_match=thehash)
+        response = self.client.put(path, profile, content_type=self.content_type, if_match=thehash, HTTP_AUTHORIZATION=self.auth)
         self.assertEqual(response.status_code, 412)
         self.assertIn('No resources matched', response.content)
 
@@ -149,7 +152,7 @@ class ActivityProfileTests(TestCase):
         params = {"profileId": 'http://etag.nomatch.good', "activityId": self.test_activityId1}
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(params))
         profile = {"test":"good - trying to put new profile w/ if none match etag header","obj":{"activity":"test"}}
-        response = self.client.put(path, profile, content_type=self.content_type, if_none_match='*')
+        response = self.client.put(path, profile, content_type=self.content_type, if_none_match='*', HTTP_AUTHORIZATION=self.auth)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.content, 'Success -- activity profile - method = PUT - profileId = %s' % 'http://etag.nomatch.good')
 
@@ -157,12 +160,12 @@ class ActivityProfileTests(TestCase):
         self.assertEqual(r.status_code, 200)
         self.assertEqual(r.content, '%s' % profile)
 
-        r = self.client.delete(reverse(views.activity_profile), params)
+        r = self.client.delete(reverse(views.activity_profile), params, HTTP_AUTHORIZATION=self.auth)
 
     def test_put_etag_if_none_match_bad(self):
         path = '%s?%s' % (reverse(views.activity_profile), urllib.urlencode(self.testparams1))
         profile = {"test":"error - trying to put new profile w/ if none match etag but one exists","obj":{"activity":"test"}}
-        response = self.client.put(path, profile, content_type=self.content_type, if_none_match='*')
+        response = self.client.put(path, profile, content_type=self.content_type, if_none_match='*', HTTP_AUTHORIZATION=self.auth)
         self.assertEqual(response.status_code, 412)
         self.assertEqual(response.content, 'Resource detected')
 
@@ -179,7 +182,7 @@ class ActivityProfileTests(TestCase):
         #self.assertEqual(response['etag'], '"%s"' % resp_hash)
         params = {'activityId': self.test_activityId2, 'profileId': self.testprofileId2}
 
-        self.client.delete(reverse(views.activity_profile), params)
+        self.client.delete(reverse(views.activity_profile), params, HTTP_AUTHORIZATION=self.auth)
 
     def test_get_activity_profileId(self):
         response = self.client.get(reverse(views.activity_profile), {'activityId':self.test_activityId1,'profileId':self.testprofileId1})
@@ -189,7 +192,7 @@ class ActivityProfileTests(TestCase):
         self.assertEqual(response['etag'], '"%s"' % resp_hash)    
         params = {'activityId': self.test_activityId1, 'profileId': self.testprofileId1}
 
-        self.client.delete(reverse(views.activity_profile), params)
+        self.client.delete(reverse(views.activity_profile), params, HTTP_AUTHORIZATION=self.auth)
 
 
     def test_get_activity_since(self):
@@ -203,7 +206,7 @@ class ActivityProfileTests(TestCase):
         #self.assertEqual(response['etag'], '"%s"' % resp_hash)
         params = {'activityId': self.test_activityId3, 'profileId': self.testprofileId3}
 
-        self.client.delete(reverse(views.activity_profile), params)
+        self.client.delete(reverse(views.activity_profile), params, HTTP_AUTHORIZATION=self.auth)
     
     def test_get_no_activity_profileId(self):
         response = self.client.get(reverse(views.activity_profile), {'profileId': self.testprofileId3})
@@ -217,6 +220,6 @@ class ActivityProfileTests(TestCase):
         self.assertEqual(response.content, 'Error -- activity_profile - method = GET, but no activityId parameter.. the activityId parameter is required')
     
     def test_delete(self):
-        response = self.client.delete(reverse(views.activity_profile), {'activityId':self.other_activityId, 'profileId':self.otherprofileId1})
+        response = self.client.delete(reverse(views.activity_profile), {'activityId':self.other_activityId, 'profileId':self.otherprofileId1}, HTTP_AUTHORIZATION=self.auth)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.content, 'Success -- activity profile - method = DELETE - profileId = %s' % self.otherprofileId1)
